Heroku

+++ src/main/java/org/generation/blogPessoal/controller/PostagemController.java
@@ -3,6 +3,8 @@ package org.generation.blogPessoal.controller;
 import java.util.List;
 import java.util.Optional;
 
+import javax.validation.Valid;
+
 import org.generation.blogPessoal.model.Postagem;
 import org.generation.blogPessoal.repository.PostagemRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,40 +18,61 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
-@RequestMapping ("/Postagem")
+@RequestMapping("/postagens")
 @CrossOrigin("*")
-
 public class PostagemController {
+	@Autowired
+	private PostagemRepository repository;
 	
-		@Autowired
-		private PostagemRepository repository;
+	@GetMapping
+	public ResponseEntity<List<Postagem>> GetAll(){
 		
-		@GetMapping
-		public ResponseEntity<List<Postagem>> findAllPostagens(){
-			return ResponseEntity.ok(repository.findAll());
-		}
+		return ResponseEntity.ok(repository.findAll());
+	}
+	
+	
+	@GetMapping("/{id}")
+	public ResponseEntity<Postagem> GetById(@PathVariable long id){
+		return repository.findById(id).map(resp -> ResponseEntity.ok(resp)).orElse(ResponseEntity.notFound().build());
+	}
+	
+	
+	@GetMapping("/titulo/{titulo}")
+	public ResponseEntity<List<Postagem>>GetByTitulo(@PathVariable String titulo){
 		
-		@GetMapping("/porTitulo/{titulo}")
-		public ResponseEntity<Postagem> getByTitulo(@PathVariable("titulo") String titulo){
-			return ResponseEntity.ok(repository.getByTituloContainingIgnoreCase(titulo));
-		}
+			return ResponseEntity.ok(repository.findAllByTituloContainingIgnoreCase(titulo));
+	}
+	
+	@PostMapping
+	public ResponseEntity<Postagem> post (@Valid @RequestBody Postagem postagem){
+		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
+	}
+	
+	@PutMapping
+	public ResponseEntity<Postagem> put (@Valid @RequestBody Postagem postagem){
+		  return repository.findById(postagem.getId())
+
+			        .map(resp -> ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem)))
+
+			        .orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
+	}
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	@DeleteMapping("/{id}")
+	
+	public void delete(@PathVariable long id) {
+		Optional<Postagem> post = repository.findById(id);
+
+        if(post.isEmpty())
+
+	        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
+        repository.deleteById(id);
 		
-		@PostMapping
-		public ResponseEntity<Postagem> post (@RequestBody Postagem postagem){
-			return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
-		}
-		
-		@PutMapping
-		public ResponseEntity<Postagem> put (@RequestBody Postagem postagem){
-			return ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem));
-		}
-		
-		@DeleteMapping("/{id}")
-		public void delete(@PathVariable Long id) {
-			repository.deleteById(id);
-		}
-		
+	}
+	
 }
