+++ src/main/java/org/generation/blogPessoal/controller/TemaController.java
@@ -1,6 +1,9 @@
 package org.generation.blogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.generation.blogPessoal.model.Tema;
 import org.generation.blogPessoal.repository.TemaRepository;
@@ -16,10 +19,12 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
+
 
 @RestController
-@CrossOrigin (origins = "*", allowedHeaders = "*")
-@RequestMapping ("/tema")
+@CrossOrigin(origins = "*", allowedHeaders = "*")
+@RequestMapping("/tema")
 public class TemaController {
 	
 	@Autowired
@@ -27,31 +32,46 @@ public class TemaController {
 	
 	@GetMapping
 	public ResponseEntity<List<Tema>> getAll(){
+		
 		return ResponseEntity.ok(repository.findAll());
 	}
 	
+	
 	@GetMapping("/{id}")
-	public ResponseEntity<Tema> getById(@PathVariable("id") Long id){
-		return repository.findById(id).map(resp -> ResponseEntity.ok(resp)).orElse(ResponseEntity.notFound().build());
+	public ResponseEntity<Tema> getById(@PathVariable long id){
+		return repository.findById(id).map(resp -> ResponseEntity.ok(resp))
+				.orElse(ResponseEntity.notFound().build());
 	}
 	
 	@GetMapping("/nome/{nome}")
-	public ResponseEntity<List<Tema>> getByName(@PathVariable("nome") String nome){
+	public ResponseEntity<List<Tema>> getByName(@PathVariable String nome){
 		return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(nome));
 	}
 	
 	@PostMapping
-	public ResponseEntity<Tema> post(@RequestBody Tema tema){
+	public ResponseEntity<Tema> post (@Valid @RequestBody Tema tema){
 		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema));
 	}
 	
 	@PutMapping
-	public ResponseEntity<Tema> put(@RequestBody Tema tema){
-		return ResponseEntity.ok(repository.save(tema));
+	public ResponseEntity<Tema> put (@Valid @RequestBody Tema tema){
+		return repository.findById(tema.getId())
+
+		        .map(resp -> ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema)))
+
+		        .orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
 	
 	@DeleteMapping("/{id}")
-	public void delete(@PathVariable("id") Long id) {
-		repository.deleteById(id);
+	public void delete(@PathVariable long id) {
+		Optional<Tema> tema = repository.findById(id);
+
+        if(tema.isEmpty())
+
+	        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
+        repository.deleteById(id);
 	}
+	
+	
 }
